Moved YfinanceTicker and AlpacaTicker to tickers file
