+++ qf_lib/common/tickers/tickers.py
@@ -321,3 +321,29 @@ class QuandlTicker(Ticker):
             return to_ticker(ticker_str)
         else:
             return [to_ticker(t) for t in ticker_str]
+
+
+class YFinanceTicker(Ticker):
+    def __init__(self, ticker: str, security_type: SecurityType = SecurityType.STOCK,
+                 point_value: int = 1, currency: Optional[str] = None):
+        super().__init__(ticker, security_type, point_value, currency)
+
+    @classmethod
+    def from_string(self, ticker_str: Union[str, Sequence[str]]) -> Union['Ticker', Sequence['Ticker']]:
+        if isinstance(ticker_str, str):
+            return YFinanceTicker(ticker_str)
+        else:
+            return [YFinanceTicker(t) for t in ticker_str]
+
+
+class AlpacaTicker(Ticker):
+    def __init__(self, ticker: str, security_type: SecurityType = SecurityType.STOCK,
+                 point_value: int = 1, currency: Optional[str] = None):
+        super().__init__(ticker, security_type, point_value, currency)
+
+    @classmethod
+    def from_string(self, ticker_str: Union[str, Sequence[str]]) -> Union['Ticker', Sequence['Ticker']]:
+        if isinstance(ticker_str, str):
+            return AlpacaTicker(ticker_str)
+        else:
+            return [AlpacaTicker(t) for t in ticker_str]
