+++ qf_lib/data_providers/yfinance/yfinance_data_provider.py
@@ -19,7 +19,7 @@ from pandas import MultiIndex
 
 from qf_lib.common.enums.frequency import Frequency
 from qf_lib.common.enums.price_field import PriceField
-from qf_lib.common.tickers.tickers import Ticker
+from qf_lib.common.tickers.tickers import Ticker, YFinanceTicker
 from qf_lib.common.utils.dateutils.relative_delta import RelativeDelta
 from qf_lib.common.utils.dateutils.timer import Timer
 from qf_lib.common.utils.miscellaneous.to_list_conversion import convert_to_list
@@ -28,7 +28,6 @@ from qf_lib.containers.qf_data_array import QFDataArray
 from qf_lib.containers.series.qf_series import QFSeries
 from qf_lib.data_providers.abstract_price_data_provider import AbstractPriceDataProvider
 from qf_lib.data_providers.helpers import normalize_data_array
-from qf_lib.data_providers.yfinance.yfinance_ticker import YFinanceTicker
 
 try:
     import yfinance as yf
