+++ qf_lib/tests/integration_tests/data_providers/yfinance_data_provider/test_yfinance_data_provider.py
@@ -28,7 +28,7 @@ from qf_lib.containers.dataframe.qf_dataframe import QFDataFrame
 from qf_lib.containers.qf_data_array import QFDataArray
 from qf_lib.containers.series.qf_series import QFSeries
 from qf_lib.data_providers.yfinance.yfinance_data_provider import YFinanceDataProvider, is_yfinance_intalled
-from qf_lib.data_providers.yfinance.yfinance_ticker import YFinanceTicker
+from qf_lib.common.tickers.tickers import YFinanceTicker
 from qf_lib.tests.helpers.testing_tools.containers_comparison import assert_series_equal, assert_dataframes_equal, \
     assert_dataarrays_equal
